backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.30
psycopg[binary]>=3.1.0
pydantic>=2.10.0
pydantic-settings>=2.5.0
pydantic[email]>=2.10.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
langchain-chroma>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
chromadb>=0.4.24
pypdf>=4.0.0
redis>=5.0.1
celery>=5.3.6
python-dotenv>=1.0.1
langchain-groq>=0.1.0
groq>=0.37.0
